% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjacent_vertex_weights}
\alias{adjacent_vertex_weights}
\title{List of adjacent vertex weights from igraph object}
\usage{
adjacent_vertex_weights(edge_verts, edge_weights)
}
\arguments{
\item{edge_verts}{edge vertices of igraph graph object}

\item{edge_weights}{edge weights of igraph graph object}
}
\value{
list of adjacent vertices
}
\description{
List of adjacent vertex weights from igraph object
}
\examples{
\dontrun{
edges <- igraph::as_edgelist(conosGraph)
edge.weights <- igraph::edge.attributes(conosGraph)$weight
adjacent_vertex_weights(edges, edge.weights)
}
}
