% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{adjacent_vertices}
\alias{adjacent_vertices}
\title{List of adjacent vertices from igraph object}
\usage{
adjacent_vertices(edge_verts)
}
\arguments{
\item{edge_verts}{edge vertices of igraph graph object}
}
\value{
list of adjacent vertices
}
\description{
List of adjacent vertices from igraph object
}
\examples{
\dontrun{
edges <- igraph::as_edgelist(conosGraph)
adjacent_vertices(edges)
}
}
