% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{appendSpecificityMetricsToDE}
\alias{appendSpecificityMetricsToDE}
\title{Append specificity metrics to DE}
\usage{
appendSpecificityMetricsToDE(
  de.df,
  clusters,
  cluster.id,
  p2.counts,
  low.expression.threshold = 0,
  append.auc = FALSE
)
}
\arguments{
\item{de.df}{}

\item{low.expression.threshold}{(default=0)}

\item{append.auc}{(default=FALSE)}
}
\value{
de.df
}
\description{
Append specificity metrics to DE
}
