% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{as_factor}
\alias{as_factor}
\title{convert values to factors}
\usage{
as_factor(vals)
}
\arguments{
\item{vals}{vector of values to evaluate}
}
\value{
values and levels
}
\description{
convert values to factors
}
