% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{avg_rank}
\alias{avg_rank}
\title{Average ranks of a numeric vector (ties resolved by mid-rank)}
\usage{
avg_rank(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
numeric vector of average ranks (1-based), NA values ranked last
}
\description{
Average ranks of a numeric vector (ties resolved by mid-rank)
}
\keyword{internal}
