% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkPackageInstalled}
\alias{checkPackageInstalled}
\title{Check for installed package(s)}
\usage{
checkPackageInstalled(
  pkgs,
  details = "to run this function",
  install.help = NULL,
  bioc = FALSE,
  cran = FALSE
)
}
\arguments{
\item{pkgs}{character Package name(s)}

\item{details}{character Helper text (default = "to run this function")}

\item{install.help}{character Additional information on how to install package (default = NULL)}

\item{bioc}{logical Package(s) is/are available from Bioconductor (default = FALSE)}

\item{cran}{logical Package(s) is/are available from CRAN (default = FALSE)}
}
\description{
Check whether a package is installed and suggest how to install from either CRAN or Bioconductor
}
\examples{
checkPackageInstalled("sccore", cran = TRUE)

}
