% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{computeChebyshevCoeffs}
\alias{computeChebyshevCoeffs}
\title{Compute Chebyshev Coefficients}
\usage{
computeChebyshevCoeffs(filt, l.max, m = 30, n = m + 1)
}
\arguments{
\item{filt}{graph filter function}

\item{l.max}{Maximum eigenvalue of the graph}

\item{m}{numeric Maximum order of Chebyshev coeff to compute (default=30)}

\item{n}{numeric grid order used to compute quadrature (default=m+1)}
}
\value{
vector of Chebyshev coefficients
}
\description{
Compute Chebyshev Coefficients
}
\seealso{
Other graph smoothing: 
\code{\link{heatFilter}()},
\code{\link{smoothChebyshev}()},
\code{\link{smoothSignalOnGraph}()}
}
\concept{graph smoothing}
\keyword{internal}
