% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conosGraph}
\alias{conosGraph}
\title{Conos graph from <https://github.com/kharchenkolab/conos/blob/master/vignettes/walkthrough.md>}
\format{
An object of class \code{igraph} of length 10.
}
\usage{
conosGraph
}
\description{
Conos graph from <https://github.com/kharchenkolab/conos/blob/master/vignettes/walkthrough.md>
}
\keyword{datasets}
