% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dotPlot}
\alias{dotPlot}
\title{Dot plot for group markers}
\usage{
dotPlot(
  markers,
  count.matrix,
  cell.groups,
  verbose = TRUE,
  n.cores = 1,
  marker.colour = "black",
  cluster.colour = "black",
  text.angle = 45,
  gene.order = NULL,
  cols = c("blue", "red"),
  col.min = -2.5,
  col.max = 2.5,
  dot.min = 0,
  dot.scale = 6,
  scale.by = "radius",
  scale.min = NA,
  scale.max = NA,
  xlab = "Marker",
  ylab = "Cluster",
  ...
)
}
\arguments{
\item{markers}{Vector of gene markers to plot}

\item{count.matrix}{Merged count matrix, e.g., through conos.obj$getJointCountMatrix()}

\item{cell.groups}{Named factor containing cell groups (clusters) and cell names}

\item{marker.colour}{Character or numeric vector (default="black")}

\item{cluster.colour}{Character or numeric vector (default="black")}

\item{xlab}{X axis title (default="Marker")}

\item{ylab}{Y axis title (default="Cluster")}

\item{...}{Additional input to sccore:::plapply, see man for description.}
}
\value{
ggplot2 object
}
\description{
Dot plot adapted from Seurat:::DotPlot, see man for description.
}
