% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{embeddingColorsPlot}
\alias{embeddingColorsPlot}
\title{Set colors for embedding plot. Used primarily in embeddingPlot().}
\usage{
embeddingColorsPlot(
  plot.df,
  colors = NULL,
  groups = NULL,
  geom_point_w = ggplot2::geom_point,
  gradient.range.quantile = 1,
  color.range = "symmetric",
  legend.title = NULL,
  palette = NULL
)
}
\arguments{
\item{colors}{vector of numbers, which must be shouwn with point colors, names contain cell names (default=NULL). This argument is ignored if groups are provided.}

\item{groups}{vector of cluster labels, names contain cell names (default=NULL)}

\item{geom_point_w}{(default=ggplot2::geom_point)}

\item{gradient.range.quantile}{Winsorization quantile for the numeric colors and gene gradient (default=1)}

\item{color.range}{controls range, in which colors are estimated (default="symmetric"). Pass "all" to estimate range based on all values of "colors", pass "data" to estimate it only based on colors, presented in the embedding. Alternatively you can pass vector of length 2 with (min, max) values.}

\item{legend.title}{legend title (default=NULL)}

\item{palette}{function, which accepts number of colors and return list of colors (i.e. see 'colorRampPalette') (default=NULL)}
}
\value{
ggplot2 object
}
\description{
Set colors for embedding plot. Used primarily in embeddingPlot().
}
