% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{embeddingGroupPlot}
\alias{embeddingGroupPlot}
\title{Plotting function for cluster labels, names contain cell names. Used primarily in embeddingPlot().}
\usage{
embeddingGroupPlot(
  plot.df,
  groups,
  geom_point_w = ggplot2::geom_point,
  min.cluster.size = 0,
  mark.groups = FALSE,
  font.size = c(3, 7),
  legend.title = NULL,
  shuffle.colors = FALSE,
  palette = NULL,
  ...
)
}
\arguments{
\item{groups}{vector of cluster labels, names contain cell names (default=NULL)}

\item{geom_point_w}{(default=ggplot2::geom_point)}

\item{min.cluster.size}{labels for all groups with number of cells fewer than this parameter are considered as missed (default=0). This argument is ignored if groups aren't provided}

\item{mark.groups}{plot cluster labels above points (default=TRUE)}

\item{font.size}{font size for cluster labels (default=c(3, 7)). It can either be single number for constant font size or pair (min, max) for font size depending on cluster size}

\item{legend.title}{legend title (default=NULL)}

\item{shuffle.colors}{shuffle colors (default=FALSE)}

\item{palette}{function, which accepts number of colors and return list of colors (i.e. see 'colorRampPalette') (default=NULL)}

\item{...}{}
}
\value{
ggplot2 object

ggplot2 object
}
\description{
Plotting function for cluster labels, names contain cell names. Used primarily in embeddingPlot().
}
