% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embedding.R
\name{embedGraphUmap}
\alias{embedGraphUmap}
\title{embedGraphUmap}
\usage{
embedGraphUmap(
  graph,
  min.prob = 0.001,
  min.visited.verts = 1000,
  n.cores = 1,
  max.hitting.nn.num = 0,
  max.commute.nn.num = 0,
  min.prob.lower = 1e-07,
  n.neighbors = 40,
  n.epochs = 1000,
  spread = 15,
  min.dist = 0.001,
  return.all = FALSE,
  n.sgd.cores = n.cores,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{min.prob}{(default=1e-3)}

\item{min.visited.verts}{(default=1000)}

\item{n.cores}{(default=1)}

\item{max.hitting.nn.num}{(default=0)}

\item{max.commute.nn.num}{(default=0)}

\item{min.prob.lower}{(default=1e-7)}

\item{n.neighbors}{(default=40)}

\item{n.epochs}{(default=1000)}

\item{spread}{(default=15)}

\item{min.dist}{(default=0.001)}

\item{return.all}{(default=FALSE)}

\item{n.sgd.cores}{(default=n.cores)}

\item{verbose}{(default=TRUE)}

\item{...}{}
}
\value{
result
}
\description{
embedGraphUmap
}
