% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embedding.R
\name{embedKnnGraph}
\alias{embedKnnGraph}
\title{embedKnnGraph}
\usage{
embedKnnGraph(
  commute.times,
  n.neighbors,
  names = NULL,
  target.dims = 2,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{names}{(default=NULL)}

\item{target.dims}{(default=2)}

\item{verbose}{(default=TRUE)}

\item{...}{arguments passed to uwot::umap()}
}
\value{
result
}
\description{
embedKnnGraph
}
