% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fac2col}
\alias{fac2col}
\title{Factor to Color}
\usage{
fac2col(
  x,
  s = 1,
  v = 1,
  shuffle = FALSE,
  min.group.size = 1,
  return.details = F,
  unclassified.cell.color = "gray50",
  level.colors = NULL
)
}
\description{
a utility function to translate factor into colors
}
