% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fac2palette}
\alias{fac2palette}
\title{Encodes logic of how to handle named-vector and functional palettes}
\usage{
fac2palette(groups, palette, unclassified.cell.color = "gray50")
}
\description{
Encodes logic of how to handle named-vector and functional palettes
}
