% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{findSubcommunities}
\alias{findSubcommunities}
\title{Increase resolution for a specific set of clusters}
\usage{
findSubcommunities(
  con,
  target.clusters,
  clustering = NULL,
  groups = NULL,
  method = igraph::cluster_louvain,
  ...
)
}
\arguments{
\item{con}{conos object, from <https://github.com/kharchenkolab/conos>, "Joint analysis of heterogeneous single-cell RNA-seq dataset collections", DOI: 10.1038/s41592-019-0466-z}

\item{target.clusters}{Clusters for which the resolution should be increased}

\item{clustering}{Name of clustering in the conos object to use (default=NULL). Either 'clustering' or 'groups' must be provided.}

\item{groups}{Set of clusters to use (default=NULL). Ignored if 'clustering' is not NULL.}

\item{method}{Function, used to find communities (default=igraph::cluster_louvain)}

\item{...}{Additional params passed to the community function}
}
\value{
The input conos object "subset" with the input target.clusters, thereby resulting in a magnified view of these clusters
}
\description{
Increase resolution for a specific set of clusters
}
