% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{getClusterGraph}
\alias{getClusterGraph}
\title{Collapse vertices belonging to each cluster in a graph}
\usage{
getClusterGraph(
  graph,
  groups,
  method = "sum",
  plot = FALSE,
  node.scale = 50,
  edge.scale = 50,
  edge.alpha = 0.3,
  ...
)
}
\arguments{
\item{graph}{igraph graph object Graph to be collapsed}

\item{groups}{factor on vertices describing cluster assignment (can specify integer vertex ids, or character vertex names which will be matched)}

\item{method}{string Method to be used, either "sum" or "paga" (default="sum")}

\item{plot}{boolean Whether to show collapsed graph plot (default=FALSE)}

\item{node.scale}{(default=50)}

\item{edge.scale}{(default=50)}

\item{edge.alpha}{(default=0.3)}

\item{...}{arguments passed to collapseGraphSum()}
}
\value{
collapsed graph
}
\description{
Collapse vertices belonging to each cluster in a graph
}
