% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_embedding.R
\name{graphToAdjList}
\alias{graphToAdjList}
\title{Convert igraph graph into an adjacency list}
\usage{
graphToAdjList(graph)
}
\arguments{
\item{graph}{input igraph object}
}
\value{
adjacency list, defined by list(idx=adj.list, probabilities=probs, names=edge.list.fact$levels
}
\description{
Convert igraph graph into an adjacency list
}
