% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeCountMatrices}
\alias{mergeCountMatrices}
\title{Merge list of count matrices}
\usage{
mergeCountMatrices(cms, transposed = FALSE, ...)
}
\arguments{
\item{cms}{List of count matrices}

\item{transposed}{Indicate whether cms are transposed, e.g. cells in rows and genes in columns (default=FALSE)}

\item{...}{Parameters for 'plapply' function}
}
\value{
A matrix
}
\description{
Merge count matrices
}
