% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multi2dend}
\alias{multi2dend}
\title{Translate multilevel segmentation into a dendrogram, with the lowest level of the dendrogram listing the cells}
\usage{
multi2dend(cl, counts, deep = FALSE, dist = "cor")
}
\arguments{
\item{deep}{(default=FALSE)}

\item{dist}{(default='cor')}
}
\value{
result
}
\description{
Translate multilevel segmentation into a dendrogram, with the lowest level of the dendrogram listing the cells
}
