% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plapply}
\alias{plapply}
\title{Parallel lapply}
\usage{
plapply(
  ...,
  n.cores = 1,
  progress = F,
  mc.preschedule = TRUE,
  mc.allow.recursive = TRUE
)
}
\arguments{
\item{n.cores}{Sumber of cores to use (default=1)}

\item{progress}{Show progress bar (default=F)}
}
\description{
Parallel, optionally verbose lapply. See ?parallel::mclapply for more info.
}
