% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{propagateLabels}
\alias{propagateLabels}
\title{Estimate labeling distribution for each vertex, based on provided labels.}
\usage{
propagateLabels(graph, labels, method = "diffusion", ...)
}
\arguments{
\item{graph}{igraph graph object}

\item{labels}{vector of factor or character labels, named by cell names, used in propagateLabelsSolver() and propagateLabelsDiffusion()}

\item{method}{string Type of propagation. Either 'diffusion' or 'solver'. (default='diffusion') 'solver' gives better result
but has bad asymptotics, so it is inappropriate for datasets > 20k cells.}

\item{...}{additional arguments passed to either propagateLabelsSolver() or propagateLabelsDiffusion()}
}
\value{
matrix with distribution of label probabilities for each vertex by rows.
}
\description{
Estimate labeling distribution for each vertex, based on provided labels.
}
\examples{
propagateLabels(conosGraph, labels=cellAnnotations)

}
