% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{propagateLabelsDiffusion}
\alias{propagateLabelsDiffusion}
\title{Estimate labeling distribution for each vertex, based on provided labels using Random Walk}
\usage{
propagateLabelsDiffusion(
  graph,
  labels,
  max.iters = 100,
  diffusion.fading = 10,
  diffusion.fading.const = 0.1,
  tol = 0.025,
  fixed.initial.labels = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{graph}{igraph graph object Graph input}

\item{labels}{vector of factor or character labels, named by cell names}

\item{max.iters}{integer Maximal number of iterations. (default=100)}

\item{diffusion.fading}{(default=10.0)}

\item{diffusion.fading.const}{(default=0.1)}

\item{tol}{numeric Absolute tolerance as a stopping criteria. (default=0.025)}

\item{verbose}{boolean Verbose mode. (default=TRUE)}

\item{fixed.initial.labels:}{prohibit changes of initial labels during diffusion. (default=TRUE)}
}
\value{
matrix
}
\description{
Estimate labeling distribution for each vertex, based on provided labels using Random Walk
}
