% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{propagateLabelsSolver}
\alias{propagateLabelsSolver}
\title{Propagate labels using Zhu, Ghahramani, Lafferty (2003) algorithm, "Semi-Supervised Learning Using Gaussian Fields and Harmonic Functions" <http://mlg.eng.cam.ac.uk/zoubin/papers/zgl.pdf>}
\usage{
propagateLabelsSolver(graph, labels, solver = "mumps")
}
\arguments{
\item{graph}{igraph graph object Graph input}

\item{labels}{vector of factor or character labels, named by cell names}

\item{solver}{Method of solver to use (default="mumps"), either "Matrix" or "mumps" (i.e. "rmumps::Rmumps")}
}
\value{
result from Matrix::solve() or rmumps::Rmumps
}
\description{
Propagate labels using Zhu, Ghahramani, Lafferty (2003) algorithm, "Semi-Supervised Learning Using Gaussian Fields and Harmonic Functions" <http://mlg.eng.cam.ac.uk/zoubin/papers/zgl.pdf>
}
\examples{
\donttest{
propagateLabelsSolver(conosGraph, labels=cellAnnotations)
}
}
