% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setMinMax}
\alias{setMinMax}
\title{Set range for values in object}
\usage{
setMinMax(obj, min, max)
}
\arguments{
\item{obj}{Object to manipulate}

\item{min}{Minimum value}

\item{max}{Maximum value}
}
\value{
An object with the same dimensions as input but with altered range in values
}
\description{
Changes values outside of range to min or max. Adapted from Seurat::MinMax
}
