% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{smoothChebyshev}
\alias{smoothChebyshev}
\title{Smooth with Chebyshev Polynomials}
\usage{
smoothChebyshev(
  lap,
  coeffs,
  signal,
  l.max,
  n.cores = 1,
  progress.chunks = 5,
  progress = (progress.chunks > 1)
)
}
\arguments{
\item{lap}{graph laplacian}

\item{coeffs}{Chebyshev coefficients for a filter}

\item{signal}{signal to smooth. Matrix or vector}

\item{l.max}{maximal eigenvalue of the graph}

\item{n.cores}{number of cores for parallel run. Default: 1.}

\item{progress.chunks}{number of chunks per core for estimating progress. Large values are not suggested, as it may bring overhead. Default: 5.}

\item{progress}{flag on whether progress must be shown. Default: `progress.chunks > 1`}
}
\value{
smoothed signal
}
\description{
Smooth with Chebyshev Polynomials
}
\seealso{
Other graph smoothing: 
\code{\link{heatFilter}()},
\code{\link{smoothSignalOnGraph}()}
}
\concept{graph smoothing}
