% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{smoothSignalOnGraph}
\alias{smoothSignalOnGraph}
\title{Smooth Signal on Graph}
\usage{
smoothSignalOnGraph(signal, graph, filter, m = 50, ...)
}
\arguments{
\item{signal}{signal to be smoothed}

\item{graph}{igraph object with the graph}

\item{filter}{function that accepts signal `x` and the maximal Laplacian eigenvalue `l.max`. See \link{heatFilter} as an example.}

\item{m}{numeric Maximum order of Chebyshev coeff to compute (default=50)}

\item{...}{
  Arguments passed on to \code{\link[=smoothChebyshev]{smoothChebyshev}}
  \describe{
    \item{\code{n.cores}}{numeric Number of cores for parallel run (default=1)}
    \item{\code{progress.chunks}}{numeric Number of chunks per core for estimating progress (default=5). Large values are not suggested, as it may bring overhead.}
    \item{\code{progress}}{boolean Flag on whether progress must be shown (default=TRUE, i.e. 'progress.chunks > 1')}
  }}
}
\description{
Smooth Signal on Graph
}
\seealso{
Other graph smoothing: 
\code{\link{heatFilter}()},
\code{\link{smoothChebyshev}()}
}
\concept{graph smoothing}
