% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sn}
\alias{sn}
\title{sn setNames wrapper function}
\usage{
sn(x)
}
\arguments{
\item{x}{an object for which names attribute will be meaningful}
}
\value{
An object with names assigned equal to values
}
\description{
Wrapper function around setNames
}
