% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{styleEmbeddingPlot}
\alias{styleEmbeddingPlot}
\title{Set plot.theme, legend, ticks for embedding plot. Used primarily in embeddingPlot().}
\usage{
styleEmbeddingPlot(
  gg,
  plot.theme = NULL,
  title = NULL,
  legend.position = NULL,
  show.legend = TRUE,
  show.ticks = TRUE,
  show.labels = TRUE,
  relabel.axis = TRUE
)
}
\arguments{
\item{gg}{ggplot2 object to plot}

\item{plot.theme}{theme for the plot (default=NULL)}

\item{title}{plot title (default=NULL)}

\item{legend.position}{vector with (x, y) positions of the legend (default=NULL)}

\item{show.legend}{show legend (default=TRUE)}

\item{show.ticks}{show ticks and tick labels (default=TRUE)}

\item{show.labels}{show labels (default=TRUE)}

\item{relabel.axis}{boolean If TRUE, relabel axes with ggplot2::labs(x='Component 1', y='Component 2') (default=TRUE)}
}
\value{
ggplot2 object
}
\description{
Set plot.theme, legend, ticks for embedding plot. Used primarily in embeddingPlot().
}
