% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{val2col}
\alias{val2col}
\title{Utility function to translate values into colors.}
\usage{
val2col(x, gradientPalette = NULL, zlim = NULL, gradient.range.quantile = 0.95)
}
\arguments{
\item{x}{input values}

\item{gradientPalette}{gradient palette (default=NULL). If NULL, use colorRampPalette(c('gray90','red'), space = "Lab")(1024) if the values are non-negative; otherwise colorRampPalette(c("blue", "grey90", "red"), space = "Lab")(1024) is used}

\item{zlim}{a two-value vector specifying limits of the values that should correspond to the extremes of the color gradient}

\item{gradient.range.quantile}{extreme quantiles of values that should be trimmed prior to color mapping (default=0.95)}
}
\description{
Utility function to translate values into colors.
}
\examples{
colors <- val2col( rnorm(10) )

}
