% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_expression.R
\name{appendSpecificityMetricsToDE}
\alias{appendSpecificityMetricsToDE}
\title{Append specificity metrics to DE}
\usage{
appendSpecificityMetricsToDE(
  de.df,
  clusters,
  cluster.id,
  p2.counts,
  low.expression.threshold = 0,
  append.auc = FALSE
)
}
\arguments{
\item{de.df}{data.frame of differential expression values}

\item{clusters}{factor of clusters}

\item{cluster.id}{names of 'clusters' factor. If a cluster.id doesn't exist in cluster names, an error is thrown.}

\item{p2.counts}{counts from Pagoda2, refer to <https://github.com/kharchenkolab/pagoda2>}

\item{low.expression.threshold}{numeric Threshold to remove expression values (default=0). Values under this threshold are discarded.}

\item{append.auc}{boolean If TRUE, append AUC values (default=FALSE)}
}
\value{
data.frame of differential expression values with metrics attached
}
\description{
Append specificity metrics to DE
}
