% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{collapseGraphPaga}
\alias{collapseGraphPaga}
\title{Collapse Graph PAGA}
\usage{
collapseGraphPaga(graph, groups, linearize = T, winsorize = F)
}
\arguments{
\item{graph}{graph to be collapsed}

\item{groups}{factor on vertives describing cluster assignment (can specify integer vertex ids, or character vertex names which will be matched)}

\item{linearize}{should normally be always `TRUE`}

\item{winsorize}{winsorize final connectivity statistics value. Original PAGA has it always `TRUE`,
but in this case there is no way to distinguish level of connectivity for highly connected groups}
}
\description{
Collapse graph using PAGA 1.2 algorithm
}
