% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{collapseGraphSum}
\alias{collapseGraphSum}
\title{Collapse Graph By Sum}
\usage{
collapseGraphSum(graph, groups, normalize = TRUE)
}
\arguments{
\item{graph}{graph to be collapsed}

\item{groups}{factor on vertices describing cluster assignment (can specify integer vertex ids, or character vertex names which will be matched)}

\item{normalize}{boolean whether to recalculate edge weight as observed/expected (default=TRUE)}
}
\value{
collapsed graph
}
\description{
Collapse Graph By Sum
}
