% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extendMatrix}
\alias{extendMatrix}
\title{Extend matrix to include new columns in matrix}
\usage{
extendMatrix(mtx, col.names)
}
\arguments{
\item{mtx}{Matrix}

\item{col.names}{Columns that should be included in matrix}
}
\value{
Matrix with new columns
}
\description{
Extend matrix to include new columns in matrix
}
