% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{extendMatrix}
\alias{extendMatrix}
\title{Extend matrix to include new columns in matrix}
\usage{
extendMatrix(mtx, col.names)
}
\arguments{
\item{mtx}{Matrix}

\item{col.names}{Columns that should be included in matrix}
}
\value{
Matrix with new columns but rows retained
}
\description{
Extend matrix to include new columns in matrix
}
\examples{
library(dplyr)
gene.union <- lapply(conosClusterList, colnames) \%>\% Reduce(union, .)
extendMatrix(conosClusterList[[1]], col.names=gene.union)

}
