% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fac2col}
\alias{fac2col}
\title{Convert factor to Color}
\usage{
fac2col(
  x,
  s = 1,
  v = 1,
  shuffle = FALSE,
  min.group.size = 1,
  return.details = FALSE,
  unclassified.cell.color = "gray50",
  level.colors = NULL
)
}
\arguments{
\item{s}{(default=1)}

\item{v}{(default=1)}

\item{shuffle}{boolean (default=FALSE)}

\item{min.group.size}{(default=1)}

\item{return.details}{(default=FALSE)}

\item{unclassified.cell.color}{Color for unclassified cells (default='gray50')}

\item{level.colors}{(default=NULL)}
}
\description{
a utility function to translate a factor into colors
}
\examples{
genes = factor(c("BRAF", "NPC1", "PAX3", "BRCA2", "FMR1"))
fac2col(genes)

}
