% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{fac2palette}
\alias{fac2palette}
\title{Encodes logic of how to handle named-vector and functional palettes. Used primarily within embeddingGroupPlot()}
\usage{
fac2palette(groups, palette, unclassified.cell.color = "gray50")
}
\arguments{
\item{groups}{vector of cluster labels, names contain cell names}

\item{palette}{vector or list or function (default=NULL). Accepts number of colors and return list of colors (i.e. see 'colorRampPalette')}

\item{unclassified.cell.color}{Color for unclassified cells (default='gray50')}
}
\value{
vector or palette
}
\description{
Encodes logic of how to handle named-vector and functional palettes. Used primarily within embeddingGroupPlot()
}
