% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{jsDist}
\alias{jsDist}
\title{Jensen–Shannon distance metric (i.e. the square root of the Jensen–Shannon divergence) between the columns of a dense matrix m}
\usage{
jsDist(m, ncores = 1L)
}
\arguments{
\item{m}{Input matrix}

\item{ncores}{integer Number of cores (default=1)}
}
\value{
Vectorized version of the lower triangle as an R distance object, stats::dist()
}
\description{
Jensen–Shannon distance metric (i.e. the square root of the Jensen–Shannon divergence) between the columns of a dense matrix m
}
\examples{
ex = matrix(1:9, nrow = 3, ncol = 3)
jsDist(ex)

}
