% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{plapply}
\alias{plapply}
\title{Parallel lapply}
\usage{
plapply(
  ...,
  n.cores = 1,
  progress = FALSE,
  mc.preschedule = TRUE,
  mc.allow.recursive = TRUE
)
}
\arguments{
\item{n.cores}{Number of cores to use (default=1)}

\item{progress}{Show progress bar (default=FALSE)}
}
\description{
Parallel, optionally verbose lapply. See ?parallel::mclapply for more info.
}
\examples{

square = function(x){ x**2 }
plapply(1:10, square, n.cores=1, progress=TRUE)

}
