% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setMinMax}
\alias{setMinMax}
\title{Set range for values in object. Changes values outside of range to min or max. Adapted from Seurat::MinMax}
\usage{
setMinMax(obj, min, max)
}
\arguments{
\item{obj}{Object to manipulate}

\item{min}{Minimum value}

\item{max}{Maximum value}
}
\value{
An object with the same dimensions as input but with altered range in values
}
\description{
Set range for values in object. Changes values outside of range to min or max. Adapted from Seurat::MinMax
}
\examples{
example_matrix =  matrix(rep(c(1:5), 3), 5)
setMinMax(example_matrix, 2, 4)

}
