% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{umapEmbedding}
\alias{umapEmbedding}
\title{UMAP embedding}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12000 rows and 2 columns.
}
\usage{
umapEmbedding
}
\description{
UMAP embedding
}
\keyword{datasets}
